requirements.txt
accelerate==0.25.0
beautifulsoup4==4.12.2
boto3==1.34.3
botocore==1.34.3
bpemb==0.3.4
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
conllu==4.5.3
contourpy==1.2.0
cycler==0.12.1
Deprecated==1.2.14
filelock==3.13.1
flair==0.13.1
fonttools==4.47.0
fsspec==2023.12.2
ftfy==6.1.3
gdown==4.7.1
gensim==4.3.2
huggingface-hub==0.19.4
idna==3.6
Janome==0.5.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
more-itertools==10.1.0
mpld3==0.5.9
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pptree==3.1
protobuf==4.25.1
psutil==5.9.7
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch_revgrad==0.2.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.9.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
segtok==1.5.11
semver==3.0.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
sqlitedict==2.1.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformer-smaller-training-vocab==0.3.3
transformers==4.36.2
tweepy==4.14.0
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.2.12
Wikipedia-API==0.6.0
wrapt==1.16.0
